remove code duplication and more refactoring in conntection classes
