+++ app/src/main/java/ch/m3ts/connection/NearbyDisplayConnection.java
@@ -2,7 +2,6 @@ package ch.m3ts.connection;
 
 import android.app.AlertDialog;
 import android.content.Context;
-import android.graphics.Point;
 import android.support.annotation.NonNull;
 
 import com.google.android.gms.nearby.Nearby;
@@ -19,36 +18,24 @@ import com.google.android.gms.nearby.connection.Strategy;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import ch.m3ts.Log;
-import ch.m3ts.display.DisplayConnectCallback;
-import ch.m3ts.display.DisplayScoreEventCallback;
-import ch.m3ts.pubnub.ByteToBase64;
-import ch.m3ts.pubnub.JSONInfo;
-import ch.m3ts.tabletennis.helper.Side;
-import ch.m3ts.tabletennis.match.UICallback;
+import ch.m3ts.connection.pubnub.JSONInfo;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-public class NearbyDisplayConnection implements DisplayConnection {
+public class NearbyDisplayConnection extends ImplDisplayConnection implements DisplayConnection {
     private static final NearbyDisplayConnection instance = new NearbyDisplayConnection();
+    private final String ID = "display";
     private ConnectionLifecycleCallback connectionLifecycleCallback;
     private PayloadCallback payloadCallback;
     private Context context;
     private ConnectionsClient connection;
-    private final String ID = "display";
     private String discovererEndpointID = "";
-    private UICallback uiCallback;
-    private DisplayScoreEventCallback scoreCallback;
-    private DisplayConnectCallback pubnubConnectionCallback;
     private ConnectionCallback connectionCallback;
-    private String encodedFrameComplete;
-    private int numberOfEncodedFrameParts;
     private String endpointName;
-
 
     private NearbyDisplayConnection() {
     }
@@ -68,7 +55,12 @@ public class NearbyDisplayConnection implements DisplayConnection {
 
             @Override
             public void onPayloadReceived(@NonNull String s, @NonNull Payload payload) {
-                handleMessage(payload);
+                try {
+                    JSONObject json = new JSONObject(new String(payload.asBytes(), UTF_8));
+                    handleMessage(json);
+                } catch (JSONException ex) {
+                    Log.d("Unable to receive JSON from endpoint \n" + ex.getMessage());
+                }
             }
 
             @Override
@@ -83,7 +75,7 @@ public class NearbyDisplayConnection implements DisplayConnection {
                     connection.stopAdvertising();
                     discovererEndpointID = s;
                     connectionCallback.onConnection(endpointName);
-                    pubnubConnectionCallback.onImageTransmissionStarted(100);
+                    displayConnectCallback.onImageTransmissionStarted(100);
                 } else {
                     connectionCallback.onRejection();
                 }
@@ -131,162 +123,35 @@ public class NearbyDisplayConnection implements DisplayConnection {
                         });
     }
 
-    public void onSelectTableCorners(Point[] tableCorners) {
-        try {
-            int[] corners = new int[tableCorners.length*2];
-            for(int i = 0; i < tableCorners.length; i++) {
-                Log.d("Point"+ i+ ": "+ tableCorners[i].x + ", " + tableCorners[i].y);
-                corners[2*i] = tableCorners[i].x;
-                corners[2*i+1] = tableCorners[i].y;
-            }
-            JSONObject json = new JSONObject();
-            json.put(JSONInfo.CORNERS, new JSONArray(corners));
-            json.put(JSONInfo.EVENT_PROPERTY, "onSelectTableCorner");
-            Payload payload = Payload.fromBytes(json.toString().getBytes(UTF_8));
-            this.connection.sendPayload(this.discovererEndpointID, payload);
-        } catch (JSONException ex) {
-            Log.d("Unable to send JSON to endpoint "+this.discovererEndpointID+"\n"+ex.getMessage());
-        }
-    }
-
     public void onStartMatch(String matchType, String server) {
         try {
             JSONObject json = new JSONObject();
             json.put(JSONInfo.TYPE_PROPERTY, matchType);
             json.put(JSONInfo.SERVER_PROPERTY, server);
             json.put(JSONInfo.EVENT_PROPERTY, "onStartMatch");
-            Payload payload = Payload.fromBytes(json.toString().getBytes(UTF_8));
-            this.connection.sendPayload(this.discovererEndpointID, payload);
+            sendData(json);
         } catch (JSONException ex) {
             Log.d("Unable to send JSON to endpoint "+this.discovererEndpointID+"\n"+ex.getMessage());
         }
-    }
-
-    public void onRestartMatch() { send("onRestartMatch", null); }
-
-    public void requestStatusUpdate() {
-        send("requestStatus", null);
-    }
-
-    public void onPointDeduction(Side side) {
-        send("onPointDeduction", side.toString());
-    }
-
-    public void onPointAddition(Side side) {
-        send("onPointAddition", side.toString());
-    }
-
-    public void onRequestTableFrame() {
-        send("onRequestTableFrame", null);
-    }
-
-    public void onPause() {
-        send("onPause", null);
-    }
-
-    public void onResume() {
-        send("onResume", null);
-    }
-
-    public void setUiCallback(UICallback uiCallback) {
-        this.uiCallback = uiCallback;
-    }
-
-    public void setDisplayScoreEventCallback(DisplayScoreEventCallback displayScoreCallback) {
-        this.scoreCallback = displayScoreCallback;
-    }
-
-    public void setDisplayConnectCallback(DisplayConnectCallback displayConnectCallback) {
-        this.pubnubConnectionCallback = displayConnectCallback;
     }
 
     public void setConnectCallback(ConnectionCallback callback) {
         this.connectionCallback = callback;
     }
 
-    private void handleOnTableFrame(JSONObject json) throws JSONException {
-        int encodedFramePartIndex = json.getInt(JSONInfo.TABLE_FRAME_INDEX);
-        int numberOfFramePartsSent = json.getInt(JSONInfo.TABLE_FRAME_NUMBER_OF_PARTS);
-        String encodedFramePart = json.getString(JSONInfo.TABLE_FRAME);
-        if (encodedFramePartIndex == 0) {
-            this.numberOfEncodedFrameParts = 1;
-            this.encodedFrameComplete = encodedFramePart;
-            this.pubnubConnectionCallback.onImageTransmissionStarted(numberOfFramePartsSent);
-        } else {
-            this.numberOfEncodedFrameParts++;
-            this.encodedFrameComplete += encodedFramePart;
-            this.pubnubConnectionCallback.onImagePartReceived(encodedFramePartIndex+1);
-            if (encodedFramePartIndex == numberOfFramePartsSent-1) {
-                Log.d("number of frame parts sent: " + numberOfFramePartsSent);
-                Log.d("number of frame parts received: " + numberOfEncodedFrameParts);
-                if (this.numberOfEncodedFrameParts == numberOfFramePartsSent) {
-                    Log.d("encodedFrame length: " + this.encodedFrameComplete.length());
-                    byte[] frame = ByteToBase64.decodeToByte(this.encodedFrameComplete);
-                    Log.d("frame length: " + frame.length);
-                    this.pubnubConnectionCallback.onImageReceived(frame, json.getInt(JSONInfo.TABLE_FRAME_WIDTH), json.getInt(JSONInfo.TABLE_FRAME_HEIGHT));
-                    this.send("onAllFramePartsReceived", null);
-                } else {
-                    onRequestTableFrame();
-                }
-            }
-        }
-    }
-
-    private void send(String event, String side) {
+    protected void send(String event, String side) {
         try {
             JSONObject json = new JSONObject();
             json.put(JSONInfo.EVENT_PROPERTY, event);
             json.put(JSONInfo.SIDE_PROPERTY, side);
-            Payload payload = Payload.fromBytes(json.toString().getBytes(UTF_8));
-            this.connection.sendPayload(this.discovererEndpointID, payload);
+            sendData(json);
         } catch (JSONException ex) {
-            Log.d("Unable to send JSON to endpoint "+this.discovererEndpointID+"\n"+ex.getMessage());
+            Log.d("Unable to send JSON to endpoint " + this.discovererEndpointID + "\n" + ex.getMessage());
         }
     }
 
-    private void handleMessage(Payload payload) {
-        try {
-            JSONObject json = new JSONObject(new String(payload.asBytes(), UTF_8));
-            String event = json.getString(JSONInfo.EVENT_PROPERTY);
-            if(event != null) {
-                switch (event) {
-                    case "onMatchEnded":
-                        this.uiCallback.onMatchEnded(json.getString(JSONInfo.SIDE_PROPERTY));
-                        break;
-                    case "onScore":
-                        this.uiCallback.onScore(Side.valueOf(json.getString(JSONInfo.SIDE_PROPERTY)), Integer.parseInt(json.getString(JSONInfo.SCORE_PROPERTY)),
-                                Side.valueOf(json.getString(JSONInfo.NEXT_SERVER_PROPERTY)), Side.valueOf(json.getString(JSONInfo.LAST_SERVER_PROPERTY)));
-                        break;
-                    case "onWin":
-                        this.uiCallback.onWin(Side.valueOf(json.getString(JSONInfo.SIDE_PROPERTY)), Integer.parseInt(json.getString(JSONInfo.WINS_PROPERTY)));
-                        break;
-                    case "onReadyToServe":
-                        this.uiCallback.onReadyToServe(Side.valueOf(json.getString(JSONInfo.SIDE_PROPERTY)));
-                        break;
-                    case "onNotReadyButPlaying":
-                        this.uiCallback.onNotReadyButPlaying();
-                        break;
-                    case "onStatusUpdate":
-                        this.scoreCallback.onStatusUpdate(json.getString(JSONInfo.PLAYER_NAME_LEFT_PROPERTY), json.getString(JSONInfo.PLAYER_NAME_RIGHT_PROPERTY),
-                                Integer.parseInt(json.getString(JSONInfo.SCORE_LEFT_PROPERTY)), Integer.parseInt(json.getString(JSONInfo.SCORE_RIGHT_PROPERTY)),
-                                Integer.parseInt(json.getString(JSONInfo.WINS_LEFT_PROPERTY)), Integer.parseInt(json.getString(JSONInfo.WINS_RIGHT_PROPERTY)),
-                                Side.valueOf(json.getString(JSONInfo.NEXT_SERVER_PROPERTY)), Integer.parseInt(json.getString(JSONInfo.GAMES_NEEDED_PROPERTY)));
-                        break;
-                    case "onConnected":
-                        if(this.pubnubConnectionCallback != null) {
-                            this.pubnubConnectionCallback.onConnected();
-                        }
-                        break;
-                    case "onTableFrame":
-                        this.handleOnTableFrame(json);
-                        break;
-                    default:
-                        Log.d("Unhandled event received:\n"+json.toString());
-                        break;
-                }
-            }
-        } catch (Exception ex) {
-            Log.d("Unable to receive JSON from endpoint "+this.discovererEndpointID+"\n"+ex.getMessage());
-        }
+    protected void sendData(JSONObject json) {
+        Payload payload = Payload.fromBytes(json.toString().getBytes(UTF_8));
+        this.connection.sendPayload(this.discovererEndpointID, payload);
     }
 }
