+++ app/src/main/java/ch/m3ts/display/MatchActivity.java
@@ -18,9 +18,9 @@ import java.util.Random;
 import ch.m3ts.Log;
 import ch.m3ts.connection.DisplayConnection;
 import ch.m3ts.connection.NearbyDisplayConnection;
+import ch.m3ts.connection.pubnub.PubNubDisplayConnection;
+import ch.m3ts.connection.pubnub.PubNubFactory;
 import ch.m3ts.helper.QuitAlertDialogHelper;
-import ch.m3ts.pubnub.DisplayPubNub;
-import ch.m3ts.pubnub.PubNubFactory;
 import cz.fmo.R;
 import cz.fmo.util.Config;
 
@@ -31,7 +31,7 @@ import cz.fmo.util.Config;
  */
 @SuppressWarnings("squid:S110")
 public class MatchActivity extends FragmentActivity implements FragmentReplaceCallback {
-    private DisplayPubNub pubNub;
+    private PubNubDisplayConnection pubNub;
     private Random random = new SecureRandom();
     private AlertDialog alertDialog;
     private NearbyDisplayConnection nearbyDisplayConnection;
@@ -93,7 +93,7 @@ public class MatchActivity extends FragmentActivity implements FragmentReplaceCa
         transaction.commit();
     }
 
-    public DisplayPubNub getPubNub() {
+    public PubNubDisplayConnection getPubNub() {
         return pubNub;
     }
