+++ app/src/main/res/values/strings.xml
@@ -5,6 +5,9 @@
     <string name="play_button_text">Play</string>
     <string name="stop_button_text">Stop</string>
     <string name="errorMsgPermissions">Sorry, cannot use M3TS unless all permissions are granted</string>
+    <string name="idDisplay">m3ts_display</string>
+    <string name="idTracker">m3ts_tracker</string>
+    <string name="errorMsgNearbyConnection">Failed to request connection to the other device. Try again or switch to PubNub connection (settings).</string>
 
     <string name="videoLength">%d seconds in buffer</string>
     <string name="noVideoLength">Buffer is empty</string>
